requirements.txt
certifi==2018.10.15
Django==2.1.2
Faker==0.9.2
olefile==0.46
Pillow==5.3.0
psycopg2==2.7.5
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
